requirements.txt
streamlit>=1.28.0
tree-sitter>=0.22.0
graphviz>=0.20.1
google-generativeai>=0.3.0
# Language parsers (pip packages)
tree-sitter-python
tree-sitter-javascript
tree-sitter-java
tree-sitter-c
tree-sitter-cpp